# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Playable Replays API
  version: 0.6.0
  description: Deterministic tactical replay API with authored scenarios, guarded telemetry detection, and summary-only local retention controls. Unknown routes and unsupported methods use the structured Error schema.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /api/v1/sessions:
    post:
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                momentId: {type: string}
              required: [momentId]
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "400": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    get:
      operationId: getSession
      responses:
        "200":
          description: Current session state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/sessions/{id}/turns:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    post:
      operationId: takeTurn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                action: {$ref: "#/components/schemas/Action"}
              required: [action]
      responses:
        "200":
          description: Resolved turn
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "400": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /api/v1/sessions/{id}/reset:
    parameters:
      - name: id
        in: path
        required: true
        schema: {type: string}
    post:
      operationId: resetSession
      responses:
        "200":
          description: Reset session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "404": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches:
    get:
      operationId: listTelemetryMatches
      responses:
        "200":
          description: Ephemeral local telemetry matches, newest first
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  matches:
                    type: array
                    items: {$ref: "#/components/schemas/TelemetryMatch"}
                required: [matches]
    post:
      operationId: createTelemetryMatch
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateTelemetryMatchRequest"}
      responses:
        "201":
          description: Capturing match and ephemeral collector credential
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CreateTelemetryMatchResponse"}
        "400": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteAllTelemetryMatches
      responses:
        "200":
          description: All in-memory telemetry plus saved summaries and drafts deleted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeleteLocalDataResponse"}
  /api/v1/telemetry/matches/{id}:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
    get:
      operationId: getTelemetryMatch
      responses:
        "200":
          description: Current match and detected highlight state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryMatch"}
        "404": {$ref: "#/components/responses/Error"}
    delete:
      operationId: deleteTelemetryMatch
      responses:
        "200":
          description: Selected in-memory telemetry plus its saved summary and drafts deleted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DeleteLocalDataResponse"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/frames:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
    post:
      operationId: ingestTelemetryFrames
      security: [{CollectorToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TelemetryFrameBatch"}
      responses:
        "202":
          description: Ordered frame batch accepted and incrementally evaluated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryMatch"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/timeline:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
    get:
      operationId: getTelemetryTimeline
      responses:
        "200":
          description: Bounded identity-free movement and normalized event timeline
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryTimeline"}
        "404": {$ref: "#/components/responses/Error"}
        "410": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/finish:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
    post:
      operationId: finishTelemetryMatch
      security: [{CollectorToken: []}]
      responses:
        "200":
          description: Match finalized and provisional candidates made final
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryMatch"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/events:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
    get:
      operationId: streamTelemetryMatch
      responses:
        "200":
          description: Server-sent match snapshots using the `match` event type
          content:
            text/event-stream:
              schema: {type: string}
        "404": {$ref: "#/components/responses/Error"}
        "410": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/candidates/{candidateId}/draft:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
      - name: candidateId
        in: path
        required: true
        schema: {type: string}
    post:
      operationId: createTelemetryDraft
      responses:
        "201":
          description: Incomplete version 2.1 draft with explicit analyst completion issues
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryDraftResult"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
    get:
      operationId: getTelemetryDraft
      responses:
        "200":
          description: Current detached workbench draft and validation state
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryDraftResult"}
        "404": {$ref: "#/components/responses/Error"}
    put:
      operationId: updateTelemetryDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                scenario: {$ref: "#/components/schemas/TelemetryScenario"}
              required: [scenario]
      responses:
        "200":
          description: Locally saved draft with refreshed field and acceptance-test results
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryDraftResult"}
        "400": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/candidates/{candidateId}/draft/validate:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
      - {$ref: "#/components/parameters/TelemetryCandidateId"}
    post:
      operationId: validateTelemetryDraft
      responses:
        "200":
          description: Field validation and deterministic win/loss acceptance results
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TelemetryDraftResult"}
        "404": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/candidates/{candidateId}/draft/preview:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
      - {$ref: "#/components/parameters/TelemetryCandidateId"}
    post:
      operationId: previewTelemetryDraft
      responses:
        "201":
          description: Ephemeral local session for a fully validated draft
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DraftPreview"}
        "404": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
  /api/v1/telemetry/matches/{id}/candidates/{candidateId}/draft/review-pack:
    parameters:
      - {$ref: "#/components/parameters/TelemetryMatchId"}
      - {$ref: "#/components/parameters/TelemetryCandidateId"}
    post:
      operationId: exportTelemetryDraftReviewPack
      responses:
        "200":
          description: Separate fixture review pack; the authored library is not modified
          content:
            application/json:
              schema: {$ref: "#/components/schemas/FixtureReviewPack"}
        "404": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
...
  /api/v1/local-storage/retention:
    put:
      operationId: updateLocalStorageRetention
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                retentionDays: {type: integer, minimum: 1, maximum: 365}
              required: [retentionDays]
      responses:
        "200":
          description: Updated retention policy and post-cleanup inventory
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LocalStorageStatus"}
        "400": {$ref: "#/components/responses/Error"}
components:
  securitySchemes:
    CollectorToken:
      type: http
      scheme: bearer
      bearerFormat: ephemeral-local-token
      description: Returned once when a consented local telemetry match is created.
  parameters:
    TelemetryMatchId:
      name: id
      in: path
      required: true
      schema: {type: string}
    TelemetryCandidateId:
      name: candidateId
      in: path
      required: true
      schema: {type: string}
  schemas:
...
    Action:
      type: object
      additionalProperties: false
      properties:
        type: {type: string, enum: [move, hold, contest, retreat]}
        target: {$ref: "#/components/schemas/Point"}
      required: [type]
...
    MechanicBriefing:
      type: object
      additionalProperties: false
      properties:
        mechanics:
          type: array
          minItems: 1
          items: {$ref: "#/components/schemas/ScenarioMechanic"}
      required: [mechanics]
    Session:
      type: object
      additionalProperties: false
      properties:
        id: {type: string}
        momentId: {type: string}
        controlledUnitId: {type: string}
        scenarioGoal: {type: string}
        mechanicBriefing: {$ref: "#/components/schemas/MechanicBriefing"}
        turn: {type: integer}
        maxTurns: {type: integer}
        status: {type: string, enum: [active, won, lost]}
        outcomeReason: {type: string}
        advantage: {type: number, minimum: 0, maximum: 1, description: Rules-based state indicator; not a calibrated probability.}
        escapeProgress: {type: integer, minimum: 0}
        escapeTurnsRequired: {type: integer, minimum: 0}
        visibleEnemyCount: {type: integer, minimum: 0}
        unknownEnemyCount: {type: integer, minimum: 0}
        visionLimited: {type: boolean}
        objective: {$ref: "#/components/schemas/ObjectiveState"}
        terrain:
          type: array
          items: {$ref: "#/components/schemas/TerrainFeature"}
        lastReferenceAction: {$ref: "#/components/schemas/Action"}
        referenceReason: {type: string}
        referenceOutcomes:
          type: array
          items: {$ref: "#/components/schemas/ReferenceOutcome"}
        bestCase: {$ref: "#/components/schemas/BestCaseLine"}
        legalActions:
          type: array
          items: {type: string, enum: [move, hold, contest, retreat]}
        units:
          type: array
          items: {$ref: "#/components/schemas/Unit"}
        log:
          type: array
          items: {$ref: "#/components/schemas/LogEntry"}
        debrief:
          type: array
          items: {type: string}
      required: [id, momentId, controlledUnitId, scenarioGoal, turn, maxTurns, status, advantage, escapeProgress, escapeTurnsRequired, visibleEnemyCount, unknownEnemyCount, visionLimited, terrain, legalActions, units, log]
...
    TelemetryFrameBatch:
      type: object
      additionalProperties: false
      properties:
        schemaVersion: {type: string, const: "1.0"}
        sequence: {type: integer, minimum: 0}
        frames:
          type: array
          minItems: 1
          maxItems: 16
          items: {$ref: "#/components/schemas/LiveTelemetryFrame"}
      required: [schemaVersion, sequence, frames]
    CreateTelemetryMatchRequest:
      type: object
      additionalProperties: false
      properties:
        source: {type: string, enum: [synthetic, authorized]}
        consent: {type: boolean, const: true}
      required: [source, consent]
...
    TelemetryMatch:
      type: object
      additionalProperties: false
      properties:
        id: {type: string}
        source: {type: string, enum: [synthetic, authorized]}
        status: {type: string, enum: [capturing, finalized]}
        frameCount: {type: integer, minimum: 0}
        lastSecond: {type: integer, minimum: -1}
        expectedSequence: {type: integer, minimum: 0}
        savedLocally: {type: boolean}
        timelineAvailable: {type: boolean}
        candidates: {type: array, items: {$ref: "#/components/schemas/TelemetryCandidate"}}
      required: [id, source, status, frameCount, lastSecond, expectedSequence, savedLocally, timelineAvailable, candidates]
...
    TelemetryTimeline:
      type: object
      additionalProperties: false
      description: Bounded display trace. Source unit/team identifiers, health, gold, and credentials are intentionally omitted.
      properties:
        matchId: {type: string}
        sourceFrameCount: {type: integer, minimum: 0, maximum: 10000}
        sampleEvery: {type: integer, minimum: 1}
        truncated: {type: boolean}
        frames:
          type: array
          maxItems: 180
          items: {$ref: "#/components/schemas/TelemetryTimelineFrame"}
        events:
          type: array
          maxItems: 240
          items: {$ref: "#/components/schemas/TelemetryTimelineEvent"}
      required: [matchId, sourceFrameCount, sampleEvery, truncated, frames, events]
    CreateTelemetryMatchResponse:
      type: object
      additionalProperties: false
      properties:
        match: {$ref: "#/components/schemas/TelemetryMatch"}
        collectorToken: {type: string, minLength: 1, writeOnly: true}
      required: [match, collectorToken]
...
    DeleteLocalDataResponse:
      type: object
      additionalProperties: false
      properties:
        deletedMatches: {type: integer, minimum: 0}
        deletedDrafts: {type: integer, minimum: 0}
      required: [deletedMatches, deletedDrafts]
...
    TelemetryScenario:
      type: object
      description: Full version 2.1 scenario fixture. Draft identity, category, timestamps, score, signals, reason tags, and source detection are immutable in the workbench.
...
    TelemetryDraftResult:
      type: object
      additionalProperties: false
      properties:
        candidateId: {type: string}
        status: {type: string, enum: [incomplete, ready]}
        completionIssues: {type: array, items: {type: string}}
        fieldIssues: {type: array, items: {$ref: "#/components/schemas/DraftFieldIssue"}}
        acceptanceResults: {type: array, items: {$ref: "#/components/schemas/AcceptanceResult"}}
        canPreview: {type: boolean}
        canExport: {type: boolean}
        bundle: {$ref: "#/components/schemas/TelemetryDraftBundle"}
      required: [candidateId, status, completionIssues, fieldIssues, acceptanceResults, canPreview, canExport, bundle]
    DraftPreview:
      type: object
      additionalProperties: false
      properties:
        moment: {$ref: "#/components/schemas/MomentSummary"}
        session: {$ref: "#/components/schemas/Session"}
      required: [moment, session]
    FixtureReviewPack:
      type: object
      additionalProperties: false
      properties:
        version: {type: string, const: "2.1"}
        moments: {type: array, items: {$ref: "#/components/schemas/TelemetryScenario"}}
      required: [version, moments]
...
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}